reorganize mocks
