+++ StitchCore/Tests/StitchCoreTests/Internal/Net/MockTransport.swift
@@ -1,0 +1,9 @@
+import Foundation
+@testable import StitchCore
+
+final class MockTransport: Transport {
+    public var mockRoundTrip = FunctionMockUnitOneArg<Response, Request>()
+    internal func roundTrip(request: Request) throws -> Response {
+        return try mockRoundTrip.throwingRun(arg1: request)
+    }
+}
